Use simple vector for dominant_colors_
The number of pixels is only used inside `find_dominant_colors`,
everywhere else it is only iterated over the set to get the color.  To
make the code easier to understand (and probably a bit more efficient),
do not keep the set beyond `find_dominant_colors`, but convert to a
simple `std::vector<FaceColor>` at the end and use that everywhere else.