Use only front line pixels for line detection
Detect "front line" pixels, i.e. pixels of one colour that are close to
pixels of the neighbouring colours.  Only use these pixels for the line
detection.

This should get rid of a lot of noise and should also speed up the
classifier for getting the line, as the amount of pixels is highly
reduced.