Refactor CubeModel
Define everything with respect to generic faces 0 to 5 and then only map
colours to the faces.  This makes it easier to adapt to other cubes
where the order of colours is different.

Also change the structure a bit to make more stuff constexpr (may allow
for a bit more optimization).