test: Add test for lightblue_segmenter
Add tests that verify that the segmentation is generally working
correctly (using a test image with a ground truth segmentation) as well
as checking that providing an input image with wrong channel number
raises an error.

Test through the Python bindings as this makes writing the test easier
and it's also the way how most users will use it.