refactor: Inherit from TriCameraObservation
Instead of completely replicating the
trifinger_cameras::TriCameraObservation, inherit from it and just add
the additional fields for the object pose.

This makes the intend clearer, the code easier to maintain (assume
something is changed in the original TriCameraObservation at some point)
and also might improve compatibility (i.e. it should allow to write
functions that can work with both types of observation).