Add backend-only mode to demo_cameras
Extend demo_cameras a bit by adding a --backend-only flag.  If set, only
the back end is run, using multi-process data.
The former `--multi-process` flag has been renamed to `--frontend-only`
for consistency.

With this change, the script can be used to run the camera backend
independent of any robot-stuff.  Actually it should probably be promoted
from a demo to a full-fledged script.