Add methods for easier conversion of ObjectPose
Add methods to get the pose as an Eigen::Affine3d and the rotation part
as Eigen::Quaterniond. Further add a constructor that takes a 4x4
transformation matrix as input.

Those methods/constructor make it easier to convert to different
representations of poses (e.g. to/from matrix which can easily be
converted to OpenCV types, etc.).