Add executable manual_annotation_hdf5
Tool to manually find the ground truth pose.
It loads a frame from an HDF5 log, runs pose estimation once, to get an
initial guess and then allows the user to modify the pose via keyboard
commands.

The pose can be saved to a JSON file, which may then be used by other
applications.

Code is mostly copied from the FOPE project with modifications to adapt
it to trifinger_object_tracking (mostly for getting the initial guess
and the visualization).