HACK: Integrate fope into tricamera driver
Very hacky approach of integrating fope: Check if an environment
variable FOPE_CONFIG is set.  If yes, it is expected to hold the path
to a fope config file.  It is then used to initialize a fope estimator,
which is used for object pose estimation instead of the standard
colour-based method of this package.

**This is not meant for production** but just for quick proof of concept
testing!