Add demo_tricamera_separate_backend.py
This script shows how to connect the TriCameraObjectTrackerDriver with a
running TriCamera backend (and at the same time also serves as a test
script for this setup).

Since this will be run on the robots without graphical interface, images
are not displayed but loggers are used to save them to files.