Avoid using same camera observation twice
When using a camera frontend, keep track of the time index of the
previously used observation.  If there is no new observation since the
last iteration, wait for it.

This implementation should always use the latest camera observation, if
the camera backend runs at a higher rate than the object tracking, or
sync with the camera backend, if it runs at a lower rate.

Note: This should result in a constant rate, if the camera backend is
slower than the object tracking.  If it is faster, it will result in the
rate being defined by how long the tracking takes.