Add executable tricamera_log_to_hdf5
Can be used to convert existing log files from the native binary format
to HDF5.  Uses the same structure (in fact the same function) as
trifinger_cameras and simply extends it by the object pose information.

No compression is applied because when chunking per observation, this
actually resulted in an increase in file size.